\name{sov-package}
\alias{sov-package}
\alias{sov}
\docType{package}
\title{
\packageTitle{sov}
}
\description{
\packageDescription{sov}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sov}
\packageIndices{sov}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{sov}

Maintainer: \packageMaintainer{sov}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
