% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sov.R
\name{sov}
\alias{sov}
\title{Traditional Shapley-Owen Values}
\usage{
sov(
  estimates = NULL,
  av = NULL,
  weight_nom = FALSE,
  absolute = FALSE,
  vw = NULL,
  q = NULL,
  pr = 0.5001,
  nPoints1 = 360,
  nPoints2 = 360,
  dec = 3,
  out_dir = "output",
  print_results = FALSE
)
}
\arguments{
\item{estimates}{Estimation results from oc, wnom, or MCMCpack.}

\item{av}{Attendance vector coded attend=1, "not attend"=NA.}

\item{weight_nom}{If TRUE use weights estimated from wnom output, if FALSE use vector of 1's. ... For wnom only.}

\item{absolute}{If TRUE, q is a scalar of yeas needed to pass proposal.  If FALSE, pr is the proportion of yeas need to pass a proposal among voters attending.}

\item{vw}{Vector of weights for each voter (default 1 for each member, applied later).}

\item{q}{The quota of yeas need to pass a proposal under absolute k-majority rule (scalar, default absolute.maj(vw) applied later).}

\item{pr}{The proportion of yea votes needed to pass a proposal among voters attending for simple k-majority rule (scalar, default: 0.5001).}

\item{nPoints1}{The number of points sampled along the azimuth (the first spherical coordinate the azimuth).  Sampled from 0 to 2*pi.}

\item{nPoints2}{The number of points sampled along the inclination (the second and third spherical coordinate).  Sampled from 0 to 2*pi.}

\item{dec}{The number of decimal places reported for vs-sovs.}

\item{out_dir}{The path to the output directory. If print_results = TRUE, the default is a subdirectory of the current path called "output", which the program creates.}

\item{print_results}{If TRUE, print results to an excel file in out_dir; if FALSE, don't print results.  In both cases, results are returned.}
}
\value{
A list with data frames containing ideal points, number of pivots, name of pivots for each direction sampled, and SOVs for each voter.
}
\description{
This function calculates traditional Shapley-Owen values (SOVs) from package estimated inputs.
}
