% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_sov.R
\name{vs_sov}
\alias{vs_sov}
\title{vs_sov}
\usage{
vs_sov(
  estimates = NULL,
  weight_nom = FALSE,
  absolute = FALSE,
  vw = NULL,
  q = NULL,
  pr = 0.5001,
  votes = NULL,
  dec = 3,
  out_dir = "output",
  print_results = FALSE
)
}
\arguments{
\item{estimates}{Estimation results from oc, wnom, or mcmcPack.}

\item{weight_nom}{If TRUE use weights estimated from wnom output, if FALSE use vector of 1's. ... For wnom only.}

\item{absolute}{If TRUE, q is a scalar of yeas needed to pass proposal.  If FALSE, pr is the proportion of yeas need to pass a proposal among voters attending.}

\item{vw}{Vector of weights for each voter (default 1 for each member, applied later).}

\item{q}{The quota of yeas need to pass a proposal under absolute k-majority rule (scalar, default absolute.maj(vw) applied later ).}

\item{pr}{The proportion of yea votes needed to pass a proposal among voters attending for simple k-majority rule (scalar, default: 0.5001).}

\item{votes}{Vote matrix (legislators x rollcalls) coded yea=1, nay=0, and 9="attend but abstain".  All other values should be NA.}

\item{dec}{The number of decimal places reported for vs-sovs.}

\item{out_dir}{The path to the output directory. If print_results = TRUE, the default is a subdirectory of the current path called "output", which the program creates.}

\item{print_results}{If TRUE, print results to an excel file in out_dir; if FALSE, don't print results.  In both cases, results are returned.}
}
\value{
Calculates vote-specific shapley owen values from package estimated inputs.
}
\description{
vs_sov
}
