% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_sov_user.R
\name{vs_sov_user}
\alias{vs_sov_user}
\title{Vote-Specific Shapley-Owen Values with User Inputs}
\usage{
vs_sov_user(
  ideals = NULL,
  normals = NULL,
  midpoints = NULL,
  weight_nom = FALSE,
  absolute = FALSE,
  vw = NULL,
  q = NULL,
  pr = 0.5001,
  votes = NULL,
  dec = 3,
  out_dir = "output",
  print_results = FALSE
)
}
\arguments{
\item{ideals}{Matrix of ideal points (legislators x dimensions)}

\item{normals}{Matrix of normal vectors (rollcalls x dimensions)}

\item{midpoints}{Matrix of Midpoints (rollcalls x dimensions), i.e. intersection of the cutplane and the unobserved normal vector}

\item{weight_nom}{If TRUE use weights estimated from wnom output, if FALSE use vector of 1's. ... For wnom only.}

\item{absolute}{If TRUE, q is a scalar of yeas needed to pass proposal.  If FALSE, pr is the proportion of yeas need to pass a proposal among voters attending.}

\item{vw}{Vector of weights for each voter (default 1 for each member, applied later).}

\item{q}{The quota of yeas need to pass a proposal under absolute k-majority rule (scalar, default absolute.maj(vw) applied later ).}

\item{pr}{The proportion of yea votes needed to pass a proposal among voters attending for simple k-majority rule (scalar, default: 0.5001).}

\item{votes}{Vote matrix (legislators x rollcalls) coded yea=1, nay=0, and 9="attend but abstain".  All other values should be NA.}

\item{dec}{The number of decimal places reported for vs-sovs.}

\item{out_dir}{The path to the output directory. If print_results = TRUE, the default is a subdirectory of the current path called "output", which the program creates.}

\item{print_results}{If TRUE, print results to an excel file in out_dir; if FALSE, don't print results.  In both cases, results are returned.}
}
\value{
A list with data frames containing ideal points, vs-SOVs for each voter, number of pivots, name of pivot(s) for each roll call, and normal vectors and angles for each roll call.
}
\description{
This function calculates vote-specific Shapley-Owen Values (vs-SOVs) from data provided by the user.
}
\examples{
#############################################################################
##### ##### 1D Example.  Start with Inputs ##### #####
## --- Ideals: 3 voters in 1D -----------------------------------------------
i1 <-  0.7
i2 <-  0.0
i3 <- -0.7
ideals <- cbind(coord1D = c(i1, i2, i3))
rownames(ideals) <- paste0("i", 1:3)

## --- Normals: 2 roll calls (x+, x-) ---------------------------------------
nv1 <-  1   # points to the right
nv2 <- -1   # points to the left
normals <- cbind(dim1 = c(nv1, nv2))
rownames(normals) <- paste0("RC", 1:2)

## --- Votes: 1=yea, 0=nay, 9=attend-no-vote, NA=absent ---------------------
votes <- cbind(
  RC1 = c(1, 0, 0),   # i1 yea, i2 nay, i3 nay
  RC2 = c(0, 1, 1)    # i1 nay, i2 yea, i3 yea
)
rownames(votes) <- rownames(ideals)

## --- Equal voting weights --------------------------------------------------
vw <- rep(1, nrow(ideals))

##### EX1: Vote-specific SOV (simple majority among attendees in 1D) #####
out_simple <- vs_sov_user(
  ideals   = ideals,
  normals  = normals,
  votes    = votes,
  absolute = FALSE,    # simple k-majority
  pr       = 0.5001,   # strict majority of attendees
  vw       = vw,
  dec      = 3
)

# Aggregate results
out_simple$pivot_summary
out_simple$pivot_by_rc
out_simple$nv_and_angles


#############################################################################
##### ##### 2D Examples.  Start with Inputs ##### #####
## --- Ideals: 5 voters in 2D -----------------------------------------------
i1 <- c( 0.7,  0.7)
i2 <- c(-0.5,  0.5)
i3 <- c(-0.7, -0.7)
i4 <- c( 0.5, -0.5)
i5 <- c( 0.0,  0.0)
ideals <- rbind(i1, i2, i3, i4, i5)
rownames(ideals) <- paste0("i", 1:5)
colnames(ideals) <- c("coord1D","coord2D")

## --- Normals: 3 roll calls (x+, y+, x-) -----------------------------------
nv1 <- c( 1, 0)
nv2 <- c( 0, 1)
nv3 <- c(-1, 0)
normals <- rbind(nv1, nv2, nv3)
rownames(normals) <- paste0("RC", 1:3)

## --- Votes: 1=yea, 0=nay, 9=attend but did not vote -----------------------
votes <- cbind(
  RC1 = c(1,0,0,1,9),
  RC2 = c(1,1,0,0,0),
  RC3 = c(0,1,1,0,0)
)
rownames(votes) <- rownames(ideals)

## --- Equal voting weights ---------------------------------------------------
vw <- rep(1, nrow(ideals))

##### EX2: Simple majority (2D) -- 50\% + epsilon among attendees #####
out_simple <- vs_sov_user(
  ideals   = ideals,
  normals  = normals,
  votes    = votes,
  absolute = FALSE,
  pr       = 0.5001,
  vw       = vw,
  dec      = 3
)

out_simple$pivot_summary
out_simple$pivot_by_rc
out_simple$nv_and_angles

### Plotting (2D): one figure with ALL normals overlaid ###
if (interactive()) {
 vs_labels2d <- setNames(out_simple$pivot_summary$vs_sov, out_simple$pivot_summary$name)
 sov:::plot_sov_geometry(ideals, normals = normals, label_values = vs_labels2d, digits = 3)
}
}
